pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pet_doi_ml"
version = "0.1.0"
description = "Machine learning for Depth-of-Interaction calibration in PET imaging"
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.11"
dependencies = [
    "numpy",
    "scipy",
    "pandas",
    "scikit-learn",
    "torch",
    "xgboost",
    "matplotlib",
    "seaborn",
    "pyyaml",
    "pydantic",
    "gitpython",
    "wandb",
]

[project.optional-dependencies]
dev = ["pytest", "pytest-cov", "ruff", "mypy", "types-PyYAML", "jupyter"]

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
select = ["E", "W", "F", "I", "B"]

[tool.mypy]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "yaml"
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
